requirements.txt
fastapi~=0.115.8
uvicorn==0.22.0
sqlalchemy~=2.0.37
asyncpg==0.27.0
psycopg2-binary==2.9.10
aiofiles==23.1.0
pydantic~=2.10.6
python-dotenv==0.21.0
alembic==1.10.2
jinja2==3.1.2
app~=0.0.1
passlib~=1.7.4
python-jose~=3.3.0